requirements.txt
Django==1.5.1
jira-python==0.13
simplejson==2.3.2
parsedatetime==1.1.2
South==0.8.2
django-reversion==1.7.1
-e git://git.linaro.org/lava/django-crowd-rest-backend.git@c741f1738786c58253b400ffc27a16967702f84c#egg=django_crowd_rest_backend-dev
